requirements.txt
numpy
scipy
tqdm
lpips
mmcv
imageio
imageio-ffmpeg
opencv-python
pytorch3d
Ninja
einops
